Eliminate all GCC compiler warnings (5520 → 0)
Suppress upstream code noise via 18 CMake -Wno-* flags for patterns like
FourCC multichar constants, MSVC pragmas, and VD2 header conventions.
Fix source-level issues: VDASSERT macro producing unused-value warnings,
missing base class init in VDString copy constructors, dummy read macros
expanding to discarded expressions, enum/int ternary mismatches, unused
lambda parameter, and a format string bug passing ATDebugExpNode* to %s.

+++ src/ATIO/source/diskimage.cpp
@@ -2907,7 +2907,7 @@ void ATDiskImage::SaveATX(IVDRandomAccessStream& fs, PhysSectors& phySecs) {
 			trkhdr.mNumSectors = psecCount;
 			trkhdr.mDefSectSize = defaultSectorSize > 512 ? 3 : defaultSectorSize > 256 ? 2 : defaultSectorSize > 128 ? 1 : 0;
 			trkhdr.mDataOffset = sizeof(ATXTrackHeader);
-			trkhdr.mFlags = isTrackMFM ? ATXTrackHeader::kFlag_MFM : 0;
+			trkhdr.mFlags = isTrackMFM ? (uint32)ATXTrackHeader::kFlag_MFM : 0;
 
 			if (trackHasLongStoredSectors)
 				trkhdr.mFlags |= ATXTrackHeader::kFlag_FullSects;
